# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "A production-grade content management and review system",
        "title": "Content Review API",
        "contact": {
            "name": "API Support",
            "url": "http://www.example.com/support"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Authenticate user with email and password, returns JWT tokens",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "User login",
                "parameters": [
                    {
                        "description": "User login credentials",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "Create a new user account with email and password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "description": "User registration details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/contents": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "List content items with pagination and filtering options",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "content"
                ],
                "summary": "List all content",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Items per page",
                        "name": "per_page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Content type filter",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Content status filter",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/contents/create": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Create a new content item (MCQ, MSQ, code problem, or documentation)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "content"
                ],
                "summary": "Create new content",
                "parameters": [
                    {
                        "description": "Content creation details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateContentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
        "/reviews/pending": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Retrieve list of content pending review (Reviewer/Admin only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "review"
                ],
                "summary": "Get pending reviews",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Items per page",
                        "name": "per_page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CreateContentRequest": {
            "type": "object",
            "required": [
                "data",
                "difficulty",
                "estimated_time_minutes",
                "program_id",
                "subtopic_id",
                "topic_id",
                "type"
            ],
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.ContentData"
                },
                "difficulty": {
                    "type": "string",
                    "enum": [
                        "easy",
                        "medium",
                        "hard"
                    ]
                },
                "estimated_time_minutes": {
                    "type": "integer",
                    "maximum": 1440,
                    "minimum": 1
                },
                "program_id": {
                    "type": "string"
                },
                "subtopic_id": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "maxItems": 20,
                    "items": {
                        "type": "string"
                    }
                },
                "topic_id": {
                    "type": "string"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "question",
                        "code_problem",
                        "documentation"
                    ]
                }
            }
        },
        "dto.LoginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 8
                }
            }
        },
        "dto.RegisterRequest": {
            "type": "object",
            "required": [
                "email",
                "name",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 255
                },
                "password": {
                    "type": "string",
                    "minLength": 8
                }
            }
        },
...
        "models.CodeProblemData": {
            "type": "object",
            "properties": {
                "execution_template": {
                    "type": "string"
                },
                "solution_code": {
                    "type": "string"
                },
                "starter_code": {
                    "type": "string"
                },
                "test_cases": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.CodeTestCase"
                    }
                }
            }
        },
        "models.CodeTestCase": {
            "type": "object",
            "properties": {
                "expected_output": {
                    "type": "string"
                },
                "input": {
                    "type": "string"
                },
                "is_hidden": {
                    "type": "boolean"
                }
            }
        },
        "models.ContentData": {
            "type": "object",
            "properties": {
                "code_problem_data": {
                    "$ref": "#/definitions/models.CodeProblemData"
                },
                "correct_options": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "description": {
                    "type": "string"
                },
                "documentation_data": {
                    "$ref": "#/definitions/models.DocumentationData"
                },
                "options": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "question_text": {
                    "type": "string"
                },
                "question_type": {
                    "description": "mcq, msq, fill, short",
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "models.DocumentationData": {
            "type": "object",
            "properties": {
                "markdown_content": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header",
            "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\""
        }
    }
}